Implementing all PR comments

+++ packages/jest-snapshot/src/SnapshotFile.js
@@ -11,8 +11,7 @@ const createDirectory = require('jest-util').createDirectory;
 const fs = require('fs');
 const path = require('path');
 const prettyFormat = require('pretty-format');
-
-const SNAPSHOT_EXTENSION = '.snap';
+const SNAPSHOT_EXTENSION = 'snap';
 
 const ensureDirectoryExists = filePath => {
   try {
@@ -44,21 +43,28 @@ class SnapshotFile {
     this._uncheckedKeys = new Set(Object.keys(this._content));
   }
 
+  hasUncheckedKeys() {
+    return this._uncheckedKeys.size > 0;
+  }
+
   fileExists() {
     return fileExists(this._filename);
+  }
+
+  removeUncheckedKeys() {
+    this._uncheckedKeys.forEach(key => delete this._content[key]);
   }
 
   serialize(data) {
     return prettyFormat(data);
   }
 
-  save() {
-    const operationResult = {
+  save(updateSnapshot) {
+    const status = {
       deleted: false,
       saved: false,
     };
 
-    this._uncheckedKeys.forEach(key => delete this._content[key]);
     const isEmpty = Object.keys(this._content).length === 0;
     if ((this._dirty || this._uncheckedKeys.size) && !isEmpty) {
       const snapshots = [];
@@ -71,15 +77,17 @@ class SnapshotFile {
 
       ensureDirectoryExists(this._filename);
       fs.writeFileSync(this._filename, snapshots.join('\n\n') + '\n');
-      operationResult.saved = true;
+      status.saved = true;
     }
 
     if (isEmpty && this.fileExists()) {
-      fs.unlinkSync(this._filename);
-      operationResult.deleted = true;
+      if (updateSnapshot) {
+        fs.unlinkSync(this._filename);
+        status.deleted = true;
+      }
     }
 
-    return operationResult;
+    return status;
   }
 
   has(key) {
@@ -110,12 +118,13 @@ class SnapshotFile {
 }
 
 module.exports = {
+  SNAPSHOT_EXTENSION,
   forFile(testPath) {
     const snapshotsPath = path.join(path.dirname(testPath), '__snapshots__');
 
     const snapshotFilename = path.join(
       snapshotsPath,
-      path.basename(testPath) + SNAPSHOT_EXTENSION
+      path.basename(testPath) + '.' + SNAPSHOT_EXTENSION
     );
 
     return new SnapshotFile(snapshotFilename);
